apps/api/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
django-cors-headers
drf-spectacular==0.27.1
